+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Feat: explicit reference support `[[@name]]`.
 - Feat: add `BiMark/BiML.findTextNodes`.
 - Feat: add `BiMark/BiML.collectDefs` to get the newly collected definitions.
+- Feat: add `BiMark/BiML.collectRefs` to get the newly collected references.
 - Feat: add `BiML.BiMLCollectOptions`.
 
 ## v0.2.0